Move Distributed Lock To Internal
